refactor: move ensureVersionsDirExists function to fsUtils for better organization
